+++ packages/cadl-rlc-emitter/test/unit/generate/docs.spec.ts
@@ -161,38 +161,6 @@ describe("Doc generation testing", () => {
       `
       );
     });
-    // issue tracked https://github.com/Azure/autorest.typescript/issues/1525
-    it("should generate apiVersion description with 'api-version'", async () => {
-      const parameters = await emitParameterFromCadl(
-        `
-        @doc("The ApiVersion query parameter.")
-        model ApiVersionParameter {
-          @query
-          @doc("The API version to use for this operation.")
-          "api-version": string;
-        }
-        op test(...ApiVersionParameter): string;
-        `
-      );
-      assert.ok(parameters);
-      assertEqualContent(
-        parameters?.content!,
-        `
-      import { RequestParameters } from "@azure-rest/core-client";
-      
-      export interface TestQueryParamProperties {
-          /** The API version to use for this operation. */
-          "api-version": string;
-      }
-      
-      export interface TestQueryParam {
-          queryParameters: TestQueryParamProperties;
-      }
-      
-      export type TestParameters = TestQueryParam & RequestParameters;
-      `
-      );
-    });
   });
 
   describe("docs in clientDefinitions.ts", () => {
